moved listing.js into the projectspage folder
